+++ test/clj/rems/api/test_email.clj
@@ -3,6 +3,7 @@
             [clojure.test :refer :all]
             [rems.api.testing :refer :all]
             [rems.db.outbox :as outbox]
+            [rems.db.test-data :as test-data]
             [rems.handler :refer [handler]]
             [ring.mock.request :refer :all]))
 
@@ -10,12 +11,23 @@
   :once
   api-fixture)
 
-(deftest test-send-handler-reminder
-  (let [api-key "42"
-        user-id "developer"
-        outbox-emails (atom [])]
+(def api-key "42")
+(def user-id "developer")
 
-    (testing "sends emails"
+(defn- create-application-in-review! []
+  (let [app-id (test-data/create-application! {:actor "alice"})]
+    (test-data/command! {:type :application.command/submit
+                         :application-id app-id
+                         :actor "alice"})
+    (test-data/command! {:type :application.command/request-comment
+                         :application-id app-id
+                         :actor "developer"
+                         :commenters ["carl"]
+                         :comment ""})))
+
+(deftest test-send-handler-reminder
+  (testing "sends emails"
+    (let [outbox-emails (atom [])]
       (with-redefs [outbox/put! (fn [email]
                                   (swap! outbox-emails conj email))]
         (let [body (-> (request :post "/api/email/send-handler-reminder")
@@ -27,15 +39,61 @@
                   {:subject "Applications in progress", :to-user "handler"}]
                  (->> @outbox-emails
                       (map #(select-keys (:outbox/email %) [:subject :to-user]))
-                      (sort-by :to-user)))))))
+                      (sort-by :to-user))))))))
 
-    (testing "requires API key"
-      (let [cookie (login-with-cookies user-id)
-            csrf (get-csrf-token cookie)
-            response (-> (request :post "/api/email/send-handler-reminder")
-                         (header "Cookie" cookie)
-                         (header "x-csrf-token" csrf)
-                         handler)]
-        (is (response-is-forbidden? response))
-        (is (str/includes? (get-in response [:headers "x-rems-roles"])
-                           "logged-in"))))))
+  (testing "requires API key"
+    (let [response (-> (request :post "/api/email/send-handler-reminder")
+                       (add-login-cookies user-id)
+                       handler)]
+      (is (response-is-forbidden? response))
+      (is (logged-in? response)))))
+
+(deftest test-send-reviewer-reminder
+  (create-application-in-review!)
+
+  (testing "sends emails"
+    (let [outbox-emails (atom [])]
+      (with-redefs [outbox/put! (fn [email]
+                                  (swap! outbox-emails conj email))]
+        (let [body (-> (request :post "/api/email/send-reviewer-reminder")
+                       (authenticate api-key user-id)
+                       handler
+                       read-ok-body)]
+          (is (= "OK" body))
+          (is (= [{:subject "Applications pending review", :to-user "carl"}]
+                 (->> @outbox-emails
+                      (map #(select-keys (:outbox/email %) [:subject :to-user]))
+                      (sort-by :to-user))))))))
+
+  (testing "requires API key"
+    (let [response (-> (request :post "/api/email/send-reviewer-reminder")
+                       (add-login-cookies user-id)
+                       handler)]
+      (is (response-is-forbidden? response))
+      (is (logged-in? response)))))
+
+(deftest test-send-reminders
+  (create-application-in-review!)
+
+  (testing "sends emails"
+    (let [outbox-emails (atom [])]
+      (with-redefs [outbox/put! (fn [email]
+                                  (swap! outbox-emails conj email))]
+        (let [body (-> (request :post "/api/email/send-reminders")
+                       (authenticate api-key user-id)
+                       handler
+                       read-ok-body)]
+          (is (= "OK" body))
+          (is (= [{:subject "Applications pending review", :to-user "carl"}
+                  {:subject "Applications in progress", :to-user "developer"}
+                  {:subject "Applications in progress", :to-user "handler"}]
+                 (->> @outbox-emails
+                      (map #(select-keys (:outbox/email %) [:subject :to-user]))
+                      (sort-by :to-user))))))))
+
+  (testing "requires API key"
+    (let [response (-> (request :post "/api/email/send-reminders")
+                       (add-login-cookies user-id)
+                       handler)]
+      (is (response-is-forbidden? response))
+      (is (logged-in? response)))))