+++ CHANGELOG.md
@@ -19,9 +19,11 @@ Changes since v2.6
   the same as before. See #837
 - Applications can be exported as CSV in admin menu (#1857)
 - Added a configuration option for setting a maximum number of days for handling a new application (#1861)
-- Reminder email about open applications to all handlers. The emails can
-  be sent by calling `/api/email/send-handler-reminder` on a cron schedule
-  or similar. The API requires an API key. (#1611)
+- Added reminder emails. The emails can be sent by calling one of the following
+  APIs on a cron schedule or similar. The APIs require an API key. (#1611, #1860)
+  - `/api/email/send-handler-reminder` sends email about open applications to all handlers.
+  - `/api/email/send-reviewer-reminder` sends email about applications with open review requests to reviewers.
+  - `/api/email/send-reminders` sends all of the above emails.
 
 ### Enhancements
 - Application search tips hidden behind question mark icon (#1767)